third_party/WebKit/LayoutTests/fast/dom/shadow/shadowroot-keyframes-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS getBackgroundColor(shadowRootWithKeyframe.getElementById('box')) is "rgb(0, 0, 255)"
PASS getBackgroundColor(shadowRootWithoutKeyframe.getElementById('box')) is "rgb(255, 0, 0)"
PASS getBackgroundColor(shadowRootWithApplyAuthorStyles.getElementById('box')) is "rgb(0, 128, 0)"
